refactor: encapsulate authentication logic
remove hashing from Credential class -> migrate to servvice layer
