+++ server/src/main/java/dev/nextchat/server/auth/repository/UserRepository.java
@@ -1,6 +1,9 @@
-package dev.nextchat.server;
+package dev.nextchat.server.auth.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import dev.nextchat.server.auth.model.User;
+
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
